Move schemas => src/
