+++ src/schemas/external-event-validation-schemata.ts
